The periodic job that checks our pipeline shows inconsistency
There was a prow job failure that caused the pipeline repository to
no longer be in sync with the 2.4 release branch of this repository.

```
****
WARNING!!!  SHA mismatch in pipeline and governance-policy-spec-sync  repositories.
   pipeline: 51a7e0a694aadf646322dca4602d8ef23f096bd4
   governance-policy-spec-sync: e83761d908c0acfc9789e591c2c2cc0647115499
***
Checking prow jobs for a failure with component governance-policy-spec-sync.
****
WARNING!!!  Prow job failure: governance-policy-spec-sync 2.4.
   Link: https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/branch-ci-stolostron-governance-policy-spec-sync-release-2.4-publish/1482051467715022848
***
```

Note the failure in the link of the publish job above.  That caused
the inconsistency that was detected.  This commit just pushes a dummy
change to run the build again.

Signed-off-by: Gus Parvin <gparvin@redhat.com>